% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsl.R
\docType{package}
\name{lsl}
\alias{lsl}
\alias{lsl-package}
\title{lsl: Latent Structure Learning}
\description{
\code{lsl} is a package designed for conducting latent structure learning methods. 
In the current version, structural equation modleing 
via penalized likelihood can be implemented using the reference class 
\code{\link{lslSEM}}.
}
